pyproject.toml
[tool.poetry]
name = "google-maps-scraper"
version = "0.1.0"
description = "A tool for scraping Google Maps data."
packages = [
    { include = "google_maps_scraper", from = "src" },
]
authors = ["Ignas Šimkūnas <ignassimkunas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.4"
click = "^8.1.7"
pandas = "^2.2.2"
selenium = "^4.22.0"
webdriver-manager = "^4.0.2"
pydantic-settings = "^2.3.4"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
gspread = "^6.0.2"
google-auth = "^2.29.0" # Updated to 2.39.0 by install, keeping higher version
gspread-dataframe = "^3.3.1"
streamlit = "^1.33.0" # Added streamlit
typing-extensions = ">=4.8.0" # Explicitly add typing-extensions, required >=4.8.0 for pydantic>=2.7

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "."
ignore_missing_imports = true
follow_imports = "silent"
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_incomplete_defs = true

[tool.isort]
py_version = 311
combine_as_imports = true
profile = "black"
lines_between_types = 1
lines_after_imports = 2
